backend/requirements.txt
Django==2.2.2
pytz==2019.1
psycopg2-binary==2.7.7
sqlparse==0.3.0
djangorestframework==3.9.2